Added forbidden import rule for the business layer
