+++ source/app/business/search.py
@@ -16,16 +16,10 @@
 #  along with this program; if not, write to the Free Software Foundation,
 #  Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
-from sqlalchemy import and_
-
 from app.iris_engine.utils.tracker import track_activity
-from app.models.comments import Comments
-from app.models.cases import Cases
-from app.models.models import Client
-from app.models.iocs import Ioc
-from app.models.models import IocType
-from app.models.models import Notes
-from app.models.iocs import Tlp
+from app.datamgmt.comments import search_comments
+from app.datamgmt.case.case_notes_db import search_notes
+from app.datamgmt.case.case_iocs_db import search_iocs
 
 
 def search(search_type, search_value):
@@ -41,72 +35,3 @@ def search(search_type, search_value):
     if search_type == 'comments':
         files = search_comments(search_value)
     return files
-
-
-def search_comments(search_value):
-    search_condition = and_()
-    comments = Comments.query.filter(
-        Comments.comment_text.like(f'%{search_value}%'),
-        Cases.client_id == Client.client_id,
-        search_condition
-    ).with_entities(
-        Comments.comment_id,
-        Comments.comment_text,
-        Cases.name.label('case_name'),
-        Client.name.label('customer_name'),
-        Cases.case_id
-    ).join(
-        Comments.case
-    ).join(
-        Cases.client
-    ).order_by(
-        Client.name
-    ).all()
-
-    return [row._asdict() for row in comments]
-
-
-def search_notes(search_value):
-    search_condition = and_()
-    notes = Notes.query.filter(
-        Notes.note_content.like(f'%{search_value}%'),
-        Cases.client_id == Client.client_id,
-        search_condition
-    ).with_entities(
-        Notes.note_id,
-        Notes.note_title,
-        Cases.name.label('case_name'),
-        Client.name.label('client_name'),
-        Cases.case_id
-    ).join(
-        Notes.case
-    ).order_by(
-        Client.name
-    ).all()
-
-    return [row._asdict() for row in notes]
-
-
-def search_iocs(search_value):
-    search_condition = and_()
-    res = Ioc.query.with_entities(
-        Ioc.ioc_value.label('ioc_name'),
-        Ioc.ioc_description.label('ioc_description'),
-        Ioc.ioc_misp,
-        IocType.type_name,
-        Tlp.tlp_name,
-        Tlp.tlp_bscolor,
-        Cases.name.label('case_name'),
-        Cases.case_id,
-        Client.name.label('customer_name')
-    ).filter(
-        and_(
-            Ioc.ioc_value.like(search_value),
-            Ioc.case_id == Cases.case_id,
-            Client.client_id == Cases.client_id,
-            Ioc.ioc_tlp_id == Tlp.tlp_id,
-            search_condition
-        )
-    ).join(Ioc.ioc_type).all()
-
-    return [row._asdict() for row in res]
